REFACTOR: se agregarón validaciones en UserDTO, se agregó ErrorResponseDTO y se agregó CreateUserRequestDTO - Se mejoró el mensaje de respuesta al ocurrir una excepción de validación. - Se refactorizó el endpoint POST "users/create" para que utilice un DTO que no solicite el campo id del user. - También se agregó un gitignore para dejar de pushear los json que persisten data.
