+++ .gitignore
@@ -32,3 +32,12 @@ build/
 ### VS Code ###
 .vscode/
 .DS_Store
+
+### Persisting files ###
+src/main/resources/posts.json
+src/main/resources/products.json
+src/main/resources/users.json
+
+src/test/resources/posts.json
+src/test/resources/products.json
+src/test/resources/users.json
