+++ src/main/java/com/bootcamp/social_meli/dto/response/ErrorResponseDTO.java
@@ -1,0 +1,16 @@
+package com.bootcamp.social_meli.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.Map;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ErrorResponseDTO {
+    private int status;
+    private String message;
+    private Map<String, String> errors;
+}
