+++ src/main/java/com/bootcamp/social_meli/dto/response/ProductDTO.java
@@ -1,5 +1,7 @@
 package com.bootcamp.social_meli.dto.response;
 
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,6 +10,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductDTO {
+    @NotNull
+    @Min(1)
     private Long product_id;
     private String product_name;
     private String type;
