+++ src/main/resources/users.json
@@ -113,4 +113,14 @@
   "username" : "jorgito",
   "first_name" : "Jorge",
   "last_name" : "Calderón"
+}, {
+  "id" : 24,
+  "username" : "j",
+  "first_name" : "Jorge",
+  "last_name" : "Calderón"
+}, {
+  "id" : 25,
+  "username" : "jo",
+  "first_name" : "Jorge",
+  "last_name" : "Calderón"
 } ]
